third_party/blink/web_tests/external/wpt/svg/painting/inheritance-expected.txt

This is a testharness.js-based test.
[FAIL] Property image-rendering inherits
  assert_equals: expected "optimizequality" but got "auto"
Harness: the test ran to completion.
